Remove deprecated Numpy type spec

+++ quantiphyse_cvr/process.py
@@ -67,7 +67,7 @@ def _run_glm(worker_id, queue, data, mask, regressors, regressor_types, regresso
             shape = data_model.data_space.shape
             if data.ndim > 1:
                 shape = list(shape) + [data.shape[1]]
-            ndata = np.zeros(shape, dtype=np.float)
+            ndata = np.zeros(shape, dtype=np.float32)
             ndata[mask > 0] = data
             ret[name] = ndata
         queue.put((worker_id, data_model.data_space.size))
